platform--scheduler/requirements.txt
dnspython==2.3.0
kafka-python==2.0.2
pymongo==4.3.3
python-crontab==2.7.1
python-dateutil==2.8.2
six==1.16.0